wip on parsing files to ndjson
