+++ .pre-commit-config.yaml
@@ -9,6 +9,7 @@ repos:
     rev: 23.3.0
     hooks:
       - id: black
+        language_version: python3.11
   - repo: https://github.com/astral-sh/ruff-pre-commit
     rev: v0.0.275
     hooks:
