+++ fetcher/ticker.py
@@ -10,7 +10,9 @@ import yaml
 from prometheus_client import start_http_server
 from pydantic import parse_obj_as
 
-from fetcher.tasks import FeedConfig, fetch_feed, FeedType, KeyValue
+from fetcher.common import KeyValue, FeedConfig
+from fetcher.feed_types import FeedType
+from fetcher.tasks import fetch_feed
 
 
 def configs_to_urls(configs: List[FeedConfig]) -> List[Tuple[FeedConfig, List[KeyValue]]]:
